requirements.txt
user_agents
pymongo
dnspython
flask
python-dotenv